==> R scripts/2-HMM on cleaned data.Rmd <==
---
title: "HMM on cleaned data"
author: "Aurore Ponchon"
date: "16/10/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,warning=FALSE,message=FALSE)
```

## Fitting a HMM for each year

First step is to test different sets of parameters for length steps and angles. 
The code is commented so that it does not run but can be reproducible with the set.seed.
Best set of parameters are determined based on the highest negative lihelihood.

### Fitting a HMM on 2015 data

```{r HMM test 2015}
load("Data/NewlyCreatedData/clean_interp_loc_2015.RData")

library(moveHMM)

essai.2015<-loc.interpolated.2015
#get a travelID by day
essai.2015$ID<-essai.2015$TravelID

proc.essai<-prepData(essai.2015,type="LL",
                     coordNames=c("Longitude","Latitude")) #step in km if LL, m if UTM

par(mfrow=c(1,2))
hist(proc.essai$step)
hist(proc.essai$angle)

#finding the right set of parameters for step length and turning angles  for 3 states
 set.seed(12345)
 nruns<-25
 mles<-vector("numeric",length=nruns)

 #test for von mise distribution for angle concentrations
# for (foo in 1:nruns){
#   print(foo)
#   mu0<-sort(c(runif(1,0.3,10),runif(1,1,20),runif(1,5,30)) ) # means for step length
#   sigma0<-c(runif(1,0.3,10),runif(1,0.3,30),runif(1,0.3,10))  #sd for step length
#   zeromass0<-runif(3,0,0.01)
#   stepPar0<-c(mu0,sigma0,zeromass0)  #joint object for step length
#   angleMean0<-sample(c(0,pi),size=3,replace=T)   #mean for turning angle
#   kappa0<-runif(3,0,5)  # angle concentration
#   anglePar0<-c(angleMean0,kappa0)
# 
#   print(c(stepPar0,anglePar0))
# 
#   m0<-fitHMM(data=proc.essai,nbStates=3,stepPar0=stepPar0,anglePar0=anglePar0,
#              formula=~1,angleDist="vm")
#   mles[foo]<--m0$mod$minimum
# 
#   print( mles[foo])
# }
# 
# table(round(mles))

#lowest likelihood: -19711.51 with von mise
#parameters:  mu0<-c(2.5,4.20,17.01)
 #sigma0<-c(2.65,10.47,7.25)  
 #zeromass0<-c(0.0087,0.0025,0.0079)
 # angleMean0<-c(0,3.14,0)   
 # kappa0<-c(4.63,3.10,3.22)

 #test for wrapped cauchy for angle concentrations
#   set.seed(12345)
# for (foo in 1:nruns){
#   print(foo)
#   mu0<-sort(c(runif(1,0.3,10),runif(1,1,20),runif(1,5,30)) ) # means for step length
#   sigma0<-c(runif(1,0.3,10),runif(1,0.3,30),runif(1,0.3,10))  #sd for step length
#   zeromass0<-runif(3,0,0.01)
#   stepPar0<-c(mu0,sigma0,zeromass0)  #joint object for step length
#   angleMean0<-sample(c(0,pi),size=3,replace=T)   #mean for turning angle
#   kappa0<-runif(3,0,5)  # angle concentration
#   anglePar0<-c(angleMean0,kappa0)
# 
#   print(c(stepPar0,anglePar0))
# 
#   m0<-fitHMM(data=proc.essai,nbStates=3,stepPar0=stepPar0,anglePar0=anglePar0,
#              formula=~1,angleDist="wrpcauchy")
#   mles[foo]<--m0$mod$minimum
# 
#   print( mles[foo])
# }
# 
# table(round(mles))

 #lowest likelihood: -18638.03 with wrpcauchy
 #parameters:  mu0<-c(1.32,8.5,9.59)
 #sigma0<-c(1.42,14.50,9.51)  
 #zeromass0<-c(0.0082,0.0081,0.0091)
 # angleMean0<-c(0,0,0)  
 # kappa0<-c(0.79,0.99,0.69)

#Test with best paramaters with wrpcauchy angle distributions
mu0<-c(1.32,8.5,9.59)
sigma0<-c(1.42,14.50,9.51)  
zeromass0<-c(0.0082,0.0081,0.0091)
stepPar0<-c(mu0,sigma0,zeromass0) 
angleMean0<-c(0,0,0)  
kappa0<-c(0.79,0.99,0.69) 
anglePar0<-c(angleMean0,kappa0)
finmodwrp.2015<-fitHMM(data=proc.essai,nbStates=3,stepPar0=stepPar0,anglePar0=anglePar0,
           formula=~1,angleDist="wrpcauchy")

res<-pseudoRes(finmodwrp.2015)
par(mfrow=c(1,2))
hist(res$stepRes)
hist(res$angle)
 
plotPR(finmodwrp.2015)
AIC(finmodwrp.2015)

print(finmodwrp.2015)
plot(finmodwrp.2015,animals=4)

loc.interpolated.2015$States<-viterbi(finmodwrp.2015)
save(loc.interpolated.2015,file="Data/NewlyCreatedData/loc_interpolated_states_2015.RData")

#delimit hours falling into duty cycle on (5am to 22pm UTM)
duty.on<-c(5:21) 
loc.interpolated.day.2015 <-loc.interpolated.2015[which(as.numeric(strftime(loc.interpolated.2015$DateTime, "%H",tz="GMT")) %in% duty.on),]

save(loc.interpolated.day.2015,file="Data/NewlyCreatedData/loc_interpolated_states_day_2015.RData")
```

### Fitting HMM on 2018 data

```{r HMM test 2018}
load("Data/NewlyCreatedData/clean_interp_loc_2018.RData")


essai.2018<-loc.interpolated.2018
#get a travelID by day
essai.2018$ID<-essai.2018$TravelID

proc.essai<-prepData(essai.2018,type="LL",
                     coordNames=c("Longitude","Latitude")) #step in km if LL, m if UTM

par(mfrow=c(1,2))
hist(proc.essai$step)
hist(proc.essai$angle)


 #test for von mise distribution for angle concentrations
# for (foo in 1:nruns){
#   print(foo)
#   mu0<-sort(c(runif(1,0.3,10),runif(1,1,20),runif(1,5,30)) ) # means for step length
#   sigma0<-c(runif(1,0.3,10),runif(1,0.3,30),runif(1,0.3,10))  #sd for step length
#   stepPar0<-c(mu0,sigma0)  #joint object for step length
#   angleMean0<-sample(c(0,pi),size=3,replace=T)   #mean for turning angle
#   kappa0<-runif(3,0,5)  # angle concentration
#   anglePar0<-c(angleMean0,kappa0)
# 
#   print(c(stepPar0,anglePar0))
# 
#   m0<-fitHMM(data=proc.essai,nbStates=3,stepPar0=stepPar0,anglePar0=anglePar0,
#              formula=~1,angleDist="vm")
#   mles[foo]<--m0$mod$minimum
# 
#   print( mles[foo])
# }
# 
# table(round(mles))

#lowest likelihood: -314259 with von mise
#parameters:  mu0<-c(1.41,3.99,11.72)
 #sigma0<-c(8.63,3.43,7.98)  
 # angleMean0<-c(0,3.14,3.14)   
 # kappa0<-c(2.06,0.33,3.98)

 #test for wrapped cauchy for angle concentrations
#   set.seed(12345)
# for (foo in 1:nruns){
#   print(foo)
#   mu0<-sort(c(runif(1,0.3,7),runif(1,1,15),runif(1,5,25)) ) # means for step length
#   sigma0<-c(runif(1,0.3,10),runif(1,0.3,25),runif(1,0.3,10))  #sd for step length
#   stepPar0<-c(mu0,sigma0)  #joint object for step length
#   angleMean0<-sample(c(0,pi),size=3,replace=T)   #mean for turning angle
#   kappa0<-runif(3,0,1)  # angle concentration
#   anglePar0<-c(angleMean0,kappa0)
# 
#   print(c(stepPar0,anglePar0))
# 
#   m0<-fitHMM(data=proc.essai,nbStates=3,stepPar0=stepPar0,anglePar0=anglePar0,
#              formula=~1,angleDist="wrpcauchy")
#   mles[foo]<--m0$mod$minimum
# 
#   print( mles[foo])
# }
# 
# table(round(mles))

 #lowest likelihood: -20560.5  with wrpcauchy
 #parameters:  mu0<-c(1.28,14.42,24.80)
 #sigma0<-c(5.11,23.02,3.91)  
 # angleMean0<-c(0,3.14,3.14)  
 # kappa0<-c(0.35,0.2,0.78)


#Test with best paramaters with wrpcauchy angle distributions
mu0<-c(1.28,14.42,24.80)
 sigma0<-c(5.11,23.02,3.91)
 stepPar0<-c(mu0,sigma0)
  angleMean0<-c(0,3.14,3.14)  
 kappa0<-c(0.35,0.2,0.78)
  anglePar0<-c(angleMean0,kappa0)
finmodWPC.2018<-fitHMM(data=proc.essai,nbStates=3,stepPar0=stepPar0,anglePar0=anglePar0,
           formula=~1,angleDist="wrpcauchy")

 res<-pseudoRes(finmodWPC.2018)
 par(mfrow=c(1,2))
hist(res$stepRes)
hist(res$angle)
 
plotPR(finmodWPC.2018)
AIC(finmodWPC.2018)

print(finmodWPC.2018)
plot(finmodWPC.2018,animals=22)


loc.interpolated.2018$States<-viterbi(finmodWPC.2018)
save(loc.interpolated.2018,file="Data/NewlyCreatedData/loc_interpolated_states_2018.RData")

#delimit hours falling into duty cycle on (5am to 22pm UTM)
duty.on<-c(5:21) 
loc.interpolated.day.2018 <-loc.interpolated.2018[  which(as.numeric(strftime(loc.interpolated.2018$DateTime, 
                                                                              "%H",tz="GMT")) %in% duty.on),]


save(loc.interpolated.day.2018,file="Data/NewlyCreatedData/loc_interpolated_states_day_2018.RData")
```
