==> R scripts/3-Kernel analysis.Rmd <==
---
title: "3-Kernels"
author: "Aurore Ponchon"
date: "20/10/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,message=FALSE, warning=FALSE)
```

## Calculating kernels and bearing

### Extracting locations and paths for each selected trip for 2015

Paths are extracted based on all interpolated trips including night time data. On the contrary, points are extracted based on data for daylight only.

```{r paths and points 2015, message =FALSE, warnong=FALSE}
load("Data/NewlyCreatedData/loc_interpolated_states_2015.RData")
load("Data/NewlyCreatedData/loc_interpolated_states_day_2015.RData")

library(fields)
library(adehabitatHR)
library(maptools)
library(rgdal)
library(raster)
library(sf)
library(geosphere)
library(plyr)
library(ggplot2)
library(viridis)

#subset data for failed and successful breeders to plot paths
loc.yna.fail.2015<-subset(loc.interpolated.2015,loc.interpolated.2015$Status=="fail")
loc.yna.suc.2015<-subset(loc.interpolated.2015,loc.interpolated.2015$Status=="success")

#create shapefiles paths
sf_locs <- sf::st_as_sf(loc.yna.fail.2015, coords = c("Longitude","Latitude")) %>% 
  sf::st_set_crs(4326)

sf_lines <- sf_locs %>%
  dplyr::arrange(ID, DateTime) %>%
  sf::st_geometry() %>%
  sf::st_cast("MULTIPOINT",ids = as.integer(as.factor(sf_locs$ID))) %>%
  sf::st_cast("MULTILINESTRING") %>%
  sf::st_sf(ID = as.factor(unique(sf_locs$ID)))

st_write(sf_lines,layer="Line_fail_2015",dsn="Data/NewlyCreatedData/Shapefiles",
         driver="ESRI Shapefile", delete_layer = TRUE,append=T,quiet=T)

sf_locs <- sf::st_as_sf(loc.yna.suc.2015, coords = c("Longitude","Latitude")) %>% 
  sf::st_set_crs(4326)

sf_lines <- sf_locs %>%
  dplyr::arrange(ID, DateTime) %>%
  sf::st_geometry() %>%
  sf::st_cast("MULTIPOINT",ids = as.integer(as.factor(sf_locs$ID))) %>%
  sf::st_cast("MULTILINESTRING") %>%
  sf::st_sf(ID = as.factor(unique(sf_locs$ID)))

st_write(sf_lines,layer="Line_success_2015",dsn="Data/NewlyCreatedData/Shapefiles",
  driver="ESRI Shapefile", delete_layer = TRUE,append=T)

#subset data for failed and successful breeders for locations
loc.fail.day.2015<-subset(loc.interpolated.day.2015,loc.interpolated.day.2015$Status=="fail")
loc.fail.day.2015<-droplevels(loc.fail.day.2015)
coordinates(loc.fail.day.2015)<-c("Longitude","Latitude")
proj4string(loc.fail.day.2015) <- CRS("+proj=longlat +ellps=WGS84")
loc.fail.day.2015<-spTransform(loc.fail.day.2015, CRS("+proj=utm +zone=45 +ellps=WGS84"))
writeOGR(loc.fail.day.2015, dsn="Data/NewlyCreatedData/Shapefiles",
         layer="Locations_failed_behav_day_2015", driver="ESRI Shapefile",overwrite_layer=T)

loc.suc.day.2015<-subset(loc.interpolated.day.2015,loc.interpolated.day.2015$Status=="success")
loc.suc.day.2015<-droplevels(loc.suc.day.2015)
coordinates(loc.suc.day.2015)<-c("Longitude","Latitude")
proj4string(loc.suc.day.2015) <- CRS("+proj=longlat +ellps=WGS84")
loc.suc.day.2015<-spTransform(loc.suc.day.2015, CRS("+proj=utm +zone=45 +ellps=WGS84"))

writeOGR(loc.suc.day.2015, dsn="Data/NewlyCreatedData/Shapefiles",
         layer="Locations_success_behav_day_2015", driver="ESRI Shapefile",overwrite_layer=T)
```

### Calculation UD kernels for each group in 2015

Smoothing factor ***h*** is fixed and common to all groups both in 2015 and 2018.

```{r kernels 2015}
loc.2015<-loc.interpolated.day.2015[loc.interpolated.day.2015$States=="1"|
                                       loc.interpolated.day.2015$States=="3",]

coordinates(loc.2015)<-c("Longitude","Latitude")
proj4string(loc.2015) <- CRS("+proj=longlat +ellps=WGS84")
loc.2015<-spTransform(loc.2015, CRS("+proj=utm +zone=45 +ellps=WGS84"))

 KUD.2015<- kernelUD(loc.2015[,2],same4all=T,h=70000,grid=500)
    KUDvol.2015 <- getvolumeUD(KUD.2015)
    ver90.suc.2015<-getverticeshr(KUDvol.2015 ,90,whi="success")
    ver50.suc.2015<-getverticeshr(KUDvol.2015 ,50,whi="success")
     ver90.fail.2015<-getverticeshr(KUDvol.2015 ,90,whi="fail")
    ver50.fail.2015<-getverticeshr(KUDvol.2015 ,50,whi="fail")

    # export shapefiles
    writeOGR(ver90.suc.2015, dsn="Data/NewlyCreatedData/Shapefiles",
             layer="KUD_90_day_suc_for_2015", driver="ESRI Shapefile",overwrite_layer=T)
    writeOGR(ver50.suc.2015, dsn="Data/NewlyCreatedData/Shapefiles",
             layer="KUD_50_day_suc_for_2015", driver="ESRI Shapefile",overwrite_layer=T)
     writeOGR(ver90.fail.2015, dsn="Data/NewlyCreatedData/Shapefiles",
             layer="KUD_90_day_fail_for_2015", driver="ESRI Shapefile",overwrite_layer=T)
    writeOGR(ver50.fail.2015, dsn="Data/NewlyCreatedData/Shapefiles",
             layer="KUD_50_day_fail_for_2015", driver="ESRI Shapefile",overwrite_layer=T)
 
    #Data for habitat modelling     
 loc.dat.2015<-loc.interpolated.day.2015[loc.interpolated.day.2015$States=="1"|
                                        loc.interpolated.day.2015$States=="3",]
 
 loc.2015.fail<-subset(loc.dat.2015,loc.dat.2015$Status=="fail")
 loc.2015.suc<-subset(loc.dat.2015,loc.dat.2015$Status=="success")
  
save(loc.2015.fail,file="Data/NewlyCreatedData/ForRestLocFail2015.RData")
save(loc.2015.suc,file="Data/NewlyCreatedData/ForRestLocSuccess2015.RData")
```

### Extracting locations and paths for each selected trip for 2018

```{r paths and points 2018}
load("Data/NewlyCreatedData/loc_interpolated_states_2018.RData")
load("Data/NewlyCreatedData/loc_interpolated_states_day_2018.RData")

#subset data for failed and successful breeders to plot paths
loc.fail.2018<-subset(loc.interpolated.2018,loc.interpolated.2018$Status=="fail")
loc.suc.2018<-subset(loc.interpolated.2018,loc.interpolated.2018$Status=="success")

#create shapefiles paths
sf_locs <- sf::st_as_sf(loc.fail.2018, coords = c("Longitude","Latitude")) %>% 
  sf::st_set_crs(4326)

sf_lines <- sf_locs %>%
  dplyr::arrange(ID, DateTime) %>%
  sf::st_geometry() %>%
  sf::st_cast("MULTIPOINT",ids = as.integer(as.factor(sf_locs$ID))) %>%
  sf::st_cast("MULTILINESTRING") %>%
  sf::st_sf(ID = as.factor(unique(sf_locs$ID)))

st_write(sf_lines,layer="Line_fail_2018",dsn="Data/NewlyCreatedData/Shapefiles",
         driver="ESRI Shapefile", delete_layer = TRUE,append=T)

sf_locs <- sf::st_as_sf(loc.suc.2018, coords = c("Longitude","Latitude")) %>% 
  sf::st_set_crs(4326)

sf_lines <- sf_locs %>%
  dplyr::arrange(ID, DateTime) %>%
  sf::st_geometry() %>%
  sf::st_cast("MULTIPOINT",ids = as.integer(as.factor(sf_locs$ID))) %>%
  sf::st_cast("MULTILINESTRING") %>%
  sf::st_sf(ID = as.factor(unique(sf_locs$ID)))

st_write(sf_lines,layer="Line_success_2018",dsn="Data/NewlyCreatedData/Shapefiles",
driver="ESRI Shapefile", delete_layer = TRUE,append=T)

#subset data for failed and successful breeders for locations
loc.fail.day.2018<-subset(loc.interpolated.day.2018,loc.interpolated.day.2018$Status=="fail")
loc.fail.day.2018<-droplevels(loc.fail.day.2018)
coordinates(loc.fail.day.2018)<-c("Longitude","Latitude")
proj4string(loc.fail.day.2018) <- CRS("+proj=longlat +ellps=WGS84")
loc.fail.day.2018<-spTransform(loc.fail.day.2018, CRS("+proj=utm +zone=45 +ellps=WGS84"))
writeOGR(loc.fail.day.2018, dsn="Data/NewlyCreatedData/Shapefiles",
         layer="Locations_failed_behav_day_2018", driver="ESRI Shapefile",overwrite_layer=T)

loc.suc.day.2018<-subset(loc.interpolated.day.2018,loc.interpolated.day.2018$Status=="success")
loc.suc.day.2018<-droplevels(loc.suc.day.2018)
coordinates(loc.suc.day.2018)<-c("Longitude","Latitude")
proj4string(loc.suc.day.2018) <- CRS("+proj=longlat +ellps=WGS84")
loc.suc.day.2018<-spTransform(loc.suc.day.2018, CRS("+proj=utm +zone=45 +ellps=WGS84"))

writeOGR(loc.suc.day.2018, dsn="Data/NewlyCreatedData/Shapefiles",
         layer="Locations_success_behav_day_2018", driver="ESRI Shapefile",overwrite_layer=T)
```

### Calculating UD kernels for each group in 2018

Smoothing factor ***h*** is fixed and common to all groups both in 2015 and 2018.

```{r kernels 2018}

loc.2018<-loc.interpolated.day.2018[loc.interpolated.day.2018$States=="1"|
                                       loc.interpolated.day.2018$States=="3",]

coordinates(loc.2018)<-c("Longitude","Latitude")
proj4string(loc.2018) <- CRS("+proj=longlat +ellps=WGS84")
loc.2018<-spTransform(loc.2018, CRS("+proj=utm +zone=45 +ellps=WGS84"))

 KUD.2018<- kernelUD(loc.2018[,2],same4all=T,h=70000,grid=500)
    KUDvol.2018 <- getvolumeUD(KUD.2018)
    ver90.suc.2018<-getverticeshr(KUDvol.2018 ,90,whi="success")
    ver50.suc.2018<-getverticeshr(KUDvol.2018 ,50,whi="success")
     ver90.fail.2018<-getverticeshr(KUDvol.2018 ,90,whi="fail")
    ver50.fail.2018<-getverticeshr(KUDvol.2018 ,50,whi="fail")

    # export shapefiles
    writeOGR(ver90.suc.2018, dsn="Data/NewlyCreatedData/Shapefiles",
             layer="KUD_90_day_suc_for_2018", driver="ESRI Shapefile",overwrite_layer=T)
    writeOGR(ver50.suc.2018, dsn="Data/NewlyCreatedData/Shapefiles",
             layer="KUD_50_day_suc_for_2018", driver="ESRI Shapefile",overwrite_layer=T)
     writeOGR(ver90.fail.2018, dsn="Data/NewlyCreatedData/Shapefiles",
             layer="KUD_90_day_fail_for_2018", driver="ESRI Shapefile",overwrite_layer=T)
    writeOGR(ver50.fail.2018, dsn="Data/NewlyCreatedData/Shapefiles",
             layer="KUD_50_day_fail_for_2018", driver="ESRI Shapefile",overwrite_layer=T)

    #Data for habitat modelling
loc.dat.2018<-loc.interpolated.day.2018[loc.interpolated.day.2018$States=="1"|
                                       loc.interpolated.day.2018$States=="3",]

loc.2018.fail<-subset(loc.dat.2018,loc.dat.2018$Status=="fail")
loc.2018.suc<-subset(loc.dat.2018,loc.dat.2018$Status=="success")

save(loc.2018.fail,file="Data/NewlyCreatedData/ForRestLocFail2018.RData")
save(loc.2018.suc,file="Data/NewlyCreatedData/ForRestLocSuccess2018.RData")
```

```{r Map kernel 2015 and 2018, echo=FALSE, fig.width=15,fig.height=8}
par(mfrow=c(1,2))

plot(ver90.fail.2015,col="yellow", border="yellow",
     main="Kernel UD 2015")
plot(ver90.suc.2015,col="pink", border="pink",add=T)
plot(ver50.suc.2015,col="purple", border="purple",add=T)
plot(ver50.fail.2015,col="orange", border="orange",add=T)

plot(ver90.fail.2018,col="yellow", border="yellow",
     main="Kernel UD 2018")
plot(ver90.suc.2018,col="pink", border="pink",add=T)
plot(ver50.suc.2018,col="purple", border="purple",add=T)
plot(ver50.fail.2018,col="orange", border="orange",add=T)
```

## Calculating and plotting bearing

```{r Bearing, fig.height=7, fig.width=15, dpi=400}
load("Data/NewlyCreatedData/loc_interpolated_states_day_2015.RData")
loc.yna.2015<-loc.interpolated.day.2015

colony<-data.frame(Longitude =77.52356,Latitude=-37.8545)
loc.yna.2015$bearing<-(bearing(loc.yna.2015[,c("Longitude","Latitude")],colony,f=0))+180
loc.yna.2015$Status<-revalue(loc.yna.2015$Status,c("fail" = "Fail 2015", 
                                         "success" = "Success 2015"))
loc.yna.2015<-subset(loc.yna.2015,loc.yna.2015$States=="1"| loc.yna.2015$States=="3")
mycols<-inferno(2,begin=0.35,end=0.8)

brk<-seq(0,360,30)
g1<-ggplot(loc.yna.2015, aes(x = bearing,fill=Status)) +
  geom_histogram( color = "gray10",breaks=brk) +
  scale_x_continuous(breaks = seq(0, 360, 30),limits=c(0,360)) +
  scale_fill_manual(values = mycols) +
  coord_polar(theta="x",start = 0, direction = 1,clip="on") +
  facet_grid(. ~ Status) +
  theme_minimal() +
  theme( legend.position = "none",
    axis.text.y = element_blank(),
    axis.title = element_blank())
  
  print(g1)  

# 
#     tiff("Figures/Rose_diagram_2015.tif", width=12,height=7,res=500,units="cm",compression="lzw")
#   print(g1)
#   dev.off()

#2018  
load("Data/NewlyCreatedData/loc_interpolated_states_day_2018.RData")

loc.yna.2018<-loc.interpolated.day.2018
colony<-data.frame(Longitude =77.52356,Latitude=-37.8545)


loc.yna.2018$bearing<-(bearing(loc.yna.2018[,c("Longitude","Latitude")],colony,f=0))+180
loc.yna.2018$Status<-revalue(loc.yna.2018$Status,c("fail" = "Fail 2018", 
                                         "success" = "Success 2018"))
loc.yna.2018<-subset(loc.yna.2018,loc.yna.2018$States=="1"| loc.yna.2018$States=="3")

g2<-ggplot(loc.yna.2018, aes(x = bearing,fill=Status)) +
  geom_histogram( color = "gray10",breaks=brk) +
  scale_x_continuous(breaks = seq(0, 360, 30),limits=c(0,360)) +
  scale_fill_manual(values = mycols) +
  coord_polar(theta="x",start = 0, direction = 1,clip="on") +
  facet_grid(. ~ Status) +
  theme_minimal() +
  theme( legend.position = "none",
    axis.text.y = element_blank(),
    axis.title = element_blank())
  
  print(g2)
# 
  #   tiff("Figures/Rose_diagram_2018.tif", width=12,height=7,res=500,units="cm",compression="lzw")
  # print(g2)
  # dev.off()

```
