==> R scripts/4-Statistics on trips.Rmd <==
---
title: "4-Statistical analysis"
author: "Aurore Ponchon"
date: "20/04/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,message=FALSE,warning=FALSE,dpi=400)
```

## Statistical analysis on raw trips

```{r trip}
library(lme4)
library(lmerTest)
library(ggplot2)
library(viridis)
library(plotrix)
library(gridExtra)
library(plyr)
library(kableExtra)
library(tidyverse)
library(plotrix)
library(simr)
source("R scripts/functions.R")

load("Data/NewlyCreatedData/raw_trips_2015.RData")
load("Data/NewlyCreatedData/raw_trips_2018.RData")

rawtrips.2015<-raw_trips_summary_ind(dataset.2015)  #function from source file
rawtrips.2018<-raw_trips_summary_ind(ALL.2018) 

rawtrips.2015$Year<-2015
rawtrips.2018$Year<-2018

ALL<-rbind(rawtrips.2015,rawtrips.2018)

#Maximal distance to the colony (km)
modl1<-lmer(sqrt(Distmax) ~ Status*Year + (1|ID),data=ALL)
modl13<-lmer(sqrt(Distmax) ~ Status+Year + (1|ID),data=ALL)

anova(modl1,modl13)
anova(modl13)

#Between failed and successful breeders in 2015
var.test(ALL$Distmax[which(ALL$Status=="success")],
         ALL$Distmax[which(ALL$Status=="fail")])

powerSim(modl13)

#Duration of the trip (h)
modl2<-lmer(sqrt(TripDur) ~ Status*Year + (1|ID),data=ALL)
modl23<-lmer(sqrt(TripDur) ~ Status+Year + (1|ID),data=ALL)

anova(modl2,modl23)
anova(modl23)

powerSim(modl23)

#Between failed and successful breeders in 2015
var.test(ALL$TripDur[which(ALL$Status=="success")],
         ALL$TripDur[which(ALL$Status=="fail")])


#Total distance travelled (km)
modl3<-lmer(sqrt(TotalPath) ~ Status*Year + (1|ID),data=ALL)
modl33<-lmer(sqrt(TotalPath) ~ Status+Year + (1|ID),data=ALL)

anova(modl3,modl33)
anova(modl33)

#Between failed and successful breeders over both years
var.test(ALL$TotalPath[which(ALL$Status=="success")],
         ALL$TotalPath[which(ALL$Status=="fail")])

powerSim(modl33)
```

## Statistical analysis on nest attendance

Nest attendance is calculated as the duration between two at-sea trips.

```{r nest attendance}
load("Data/RawData/Attendance.RData")

modl4<-lmer(sqrt(Difftime) ~ Status*Year + (1|ID),data=nest)
modl41<-lmer(sqrt(Difftime) ~ Status+Year + (1|ID),data=nest)
modl42<-lmer(sqrt(Difftime) ~ Status + (1|ID),data=nest)
modl43<-lmer(sqrt(Difftime) ~ Year + (1|ID),data=nest)

anova(modl4,modl41)
anova(modl41)

powerSim(modl41)
```
```{r echo=FALSE}
nest1<-as_tibble(nest)
nest1<-nest1 %>%
  group_by(Year,Status) %>% 
  summarise(Dur=mean(Difftime),SEDur=std.error(Difftime),
            MinDur=min(Difftime),MaxDur=max(Difftime))
nest2<-as.data.frame(nest1)

kbl(nest2,caption="Summary of nest attendance") %>%
  kable_styling(bootstrap_options = c("striped", "hover"))

```

## Statistical analysis on at-sea behavioural states

```{r behaviour}
load("Data/NewlyCreatedData/loc_interpolated_states_day_2015.RData")
load("Data/NewlyCreatedData/loc_interpolated_states_day_2018.RData")

loc.interpolated.day.2015$TravelID<-paste(loc.interpolated.day.2015$TravelNb,
                           format(loc.interpolated.day.2015$DateTime,format="%d"),
                            sep=".")

behav15<-ddply(loc.interpolated.day.2015,.(ID,Status,TravelID,States),summarize,
               SumDur=sum(Difftime)/60)

trip15<-ddply(loc.interpolated.day.2015,.(ID,Status,TravelID),summarize,
              TotDur=sum(Difftime)/60)

behav15<-merge(behav15,trip15,by=c("ID","Status","TravelID"))
behav15$Prop<-behav15$SumDur/behav15$TotDur*100
behav15$Year<-2015

loc.interpolated.day.2018$TravelID<-paste(loc.interpolated.day.2018$TravelNb,
                           format(loc.interpolated.day.2018$DateTime,format="%d"),
                            sep=".")

behav18<-ddply(loc.interpolated.day.2018,.(ID,Status,TravelID,States),summarize,
               SumDur=sum(Difftime))

trip18<-ddply(loc.interpolated.day.2018,.(ID,Status,TravelID),summarize,
              TotDur=sum(Difftime))

behav18<-merge(behav18,trip18,by=c("ID","Status","TravelID"))
behav18$Prop<-behav18$SumDur/behav18$TotDur*100
behav18$Year<-2018

behav<-rbind(behav15,behav18)
behav$Year<-as.factor(behav$Year)

behav1<-ddply(behav,.(Year,Status,ID,States),summarize,MeanIndProp=mean(Prop))

behav2<-ddply(behav1,.(Year,Status,States),summarize,MeanProp=mean(MeanIndProp),
              se=std.error(MeanIndProp))

kbl(behav2,caption="Summary of at-sea behaviour") %>%
  kable_styling(bootstrap_options = c("striped", "hover"))


behav.for<-behav1[behav1$States==3,]
behav.rest<-behav1[behav1$States==1,]
behav.fly<-behav1[behav1$States==2,]

mm.for<-glmer(MeanIndProp/100~Status+Year +(1|ID),data=behav.for,family="binomial")
mm.rest<-glmer(MeanIndProp/100~Status+Year +(1|ID),data=behav.rest,family="binomial")
mm.fly<-glmer(MeanIndProp/100~Status+Year +(1|ID),data=behav.fly,family="binomial")

summary(mm.for)
summary(mm.fly)
summary(mm.rest)

```


## Plots with trip characteristics

```{r plots of trips characteristics ,echo=FALSE, fig.height=8,fig.width=15,dpi=400}
colo<-inferno(2,begin=0.35,end=0.8)
mycol<-c(colo[1], colo[2],colo[1],colo[2])

ALL$Interac<-paste(ALL$Status,ALL$Year,sep=" ")
ALL$Interac<-factor(ALL$Interac,levels=c("success 2015","fail 2015","success 2018","fail 2018"))

 #maximal distance
dist<-ggplot(ALL,aes(x=Interac,y=Distmax, fill=Interac,
                                       colour=Interac) ) +
  geom_boxplot(outlier.alpha = 0, alpha=0.3, show.legend = FALSE)+
  geom_jitter(width=0.1, show.legend = FALSE)+
  scale_y_continuous(limits=c(0,1500),breaks=seq(0,1500,500),expand=c(0.01,0.01))+
 labs(y="Maximal distance to the colony (km)",tag="a)",x="") +
  scale_fill_manual(values=mycol) +
  scale_colour_manual(values=mycol)  +
  theme(text = element_text(size=10),
        axis.text = element_text(size=10),
        legend.text = element_text(size=10),
        axis.title.x=element_blank())

#trip duration
dur<-ggplot(ALL,aes(x=Interac,y=TripDur, 
                                       fill=Interac, colour=Interac) )+
  geom_jitter(width=0.1, show.legend = FALSE)+
  geom_boxplot(outlier.alpha = 0, alpha=0.3, show.legend = FALSE)+
  labs(y="Trip duration (hrs)", x="",tag="b)") +
  scale_y_continuous(limits=c(0,400),expand=c(0.01,0.01))+
  scale_fill_manual(values=mycol) +
scale_colour_manual(values=mycol)+
  theme(text = element_text(size=10),
        axis.text = element_text(size=10),
        legend.text = element_text(size=10),
        axis.title.x=element_blank())

#total distance travelled
tot<-ggplot(ALL,aes(x=Interac,y=TotalPath, 
                                      fill=Interac, colour=Interac) )+
  geom_jitter(width=0.1, show.legend = FALSE)+
  geom_boxplot(outlier.alpha = 0, alpha=0.3, show.legend = FALSE)+
  labs(y="Total distance travelled (km)",x="",tag="c)") +
  scale_y_continuous(limits=c(0,5000),expand=c(0.01,0.01))+
  scale_fill_manual(values=mycol) +
  scale_colour_manual(values=mycol)+
  theme(text = element_text(size=10),
        axis.text = element_text(size=10),
        legend.text = element_text(size=10),
        axis.title.x=element_blank())

#nest attendance
nest$Interac<-paste(nest$Status,nest$Year,sep=" ")
nest$Interac<-factor(nest$Interac,levels=c("success 2015","fail 2015","success 2018","fail 2018"))

att<-ggplot(nest,aes(x=Interac,y=Difftime, 
                                      fill=Interac, colour=Interac) )+
  geom_jitter(width=0.1,show.legend=F)+
  geom_boxplot(outlier.alpha = 0, alpha=0.3, show.legend = FALSE)+
  labs(y="Nest attendance (hrs)",x="",tag="d)") +
  scale_y_continuous(limits=c(0,120),expand=c(0.01,0.01))+
  scale_fill_manual(values=mycol)  +
  scale_colour_manual(values=mycol)+
  theme(text = element_text(size=10),
        axis.text = element_text(size=10),
        axis.title.x=element_blank())

 # tiff("Figures/Figure 2 - Stats trip characteristics.tif",compression="lzw",
 #      res=600,width=6500,height=3500)
grid.arrange(arrangeGrob(dist,dur,tot,att,
                         ncol=2,nrow=2))

 # dev.off()
```

## At-sea behaviour

```{r plots of at-sea behaviour ,echo=FALSE, fig.height=5,fig.width=15,dpi=400}
behav$Interac<-paste(behav$Status,behav$Year,sep=" ")
behav$Interac<-factor(behav$Interac,levels=c("success 2015","fail 2015","success 2018","fail 2018"))
behav$States<-as.factor(behav$States)
behav$States<-revalue(behav$States,c("1" = "Resting",
                                   "2"="Travelling",
                                   "3" = "Foraging"))

att<-ggplot(behav,aes(x=Interac,y=Prop/100, 
                     fill=Interac, colour=Interac) )+
  geom_jitter(width=0.1, show.legend = FALSE)+
  facet_grid(.~States)+
  geom_boxplot(outlier.alpha = 0, alpha=0.5, show.legend = FALSE)+
  ylab("Proportion of time per day") +
  xlab("") +
  scale_y_continuous(limits=c(0,1),expand=c(0,0))+
  scale_colour_manual(values=mycol)+
  scale_fill_manual(values=mycol)
print(att)

```
